refactor(memory): split episodic.py and test_speaker_projection.py off the 500-line cap
Both sat at exactly 500 (scripts/checks/file_size.py --strict), so the
review round had to land the store_episode speaker_id docstring net-zero
by re-wrapping prose — the at-the-cap trap this branch already hit three
times (interaction_key.py, _facts_write.py, close_entries.py).

- episodic.py (500 → 465): the agent_state-table API — the interaction
  counter and persist/load_agent_state — moves to a new
  _EpisodicStateAPIMixin (episodic_state_api.py), the
  _EpisodicNotesAPIMixin idiom: pure delegation as a mixin so the
  methods keep their public call sites and the _ensure_db lifecycle
  gate; SQL bodies stay in episodic_queries.  The two sections are one
  concern — both live on the same agent_state row, and the upsert
  contract (persist must not reset interaction_count) only reads
  coherently with them together.  Public surface unchanged.
- test_speaker_projection.py (500 → 397): the fact-tier pins
  (TestFactProjection, TestTheFactDispatchCarriesItThrough, the
  fact_store fixture) move to test_speaker_projection_facts.py, which
  imports the shared record builders from the sibling — the
  test_close_notification_room_fan precedent.

Same suite totals before and after (4769 unit / 422 agents / 366
integration); FILEMAP regenerated for the two new modules.
