+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-08-30  
-> **2023** tracked files
+> **2025** tracked files
 
 ## Summary
 
@@ -9,8 +9,8 @@
 |---------------------|-------|
 | `internal/` | 515 |
 | `docs/` | 454 |
-| `tests/` | 436 |
-| `agents/` | 292 |
+| `tests/` | 437 |
+| `agents/` | 293 |
 | `web/` | 78 |
 | `prompts/` | 48 |
 | `cli/` | 42 |
@@ -28,7 +28,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 755 |
+| `.py` | 757 |
 | `.go` | 533 |
 | `.md` | 519 |
 | `.yaml` | 57 |
@@ -182,6 +182,7 @@
 │   │   ├── episodic_queries.py
 │   │   ├── episodic_retention.py
 │   │   ├── episodic_room_ranked.py
+│   │   ├── episodic_state_api.py
 │   │   ├── eviction.py
 │   │   ├── facade.py
 │   │   ├── facade_procedural.py
@@ -2068,6 +2069,7 @@
 │           ├── test_spawner_helpers.py
 │           ├── test_speaker_migration.py
 │           ├── test_speaker_projection.py
+│           ├── test_speaker_projection_facts.py
 │           ├── test_split_participant_migration.py
 │           ├── test_sub_agent_model_default.py
 │           ├── test_subagent_lease.py
