+++ tests/unit/python/test_speaker_projection.py
@@ -21,6 +21,10 @@ sender is not the record's speaker.  The RFC obliges this PR to
 "exclude or tag" it; it is discharged as EXCLUDE, upstream of the
 combined summarise+extract call, so no fact can be derived from another
 speaker's words and then stamped with this record's.
+
+The fact-tier half of the projection is pinned in
+:mod:`.test_speaker_projection_facts` (split at the 500-line cap); the
+record builders below are shared with it.
 """
 
 from __future__ import annotations
@@ -28,20 +32,13 @@ from __future__ import annotations
 import json
 from unittest.mock import AsyncMock, MagicMock
 
-import pytest
-
 from agents.llm_client import LLMResponse, StopReason, Usage
 from agents.memory.boundary_detectors import REASON_STRUCTURAL
-from agents.memory.facts import FactStore
 from agents.memory.interaction_tracker import InteractionTracker
 from agents.memory.interaction_types import ROOM_CLOSE_TURN_KEY, Interaction, Turn
 from agents.persona_runtime import close_path
 from agents.persona_runtime.close_entries import interaction_to_entries
 from agents.persona_runtime.episode_routing import _EpisodeRoutingMixin
-from agents.persona_runtime.fact_extractor import (
-    dispatch_facts_from_response,
-    store_extracted_facts,
-)
 from agents.persona_runtime.summarize_close import (
     summarize_closed_interaction,
 )
@@ -238,48 +235,6 @@ class TestPersistedContextAppliesTheExclusion:
         ) == ["iron-fox", "iron-fox"]
 
 
-class TestFactProjection:
-    async def test_every_tuple_in_the_batch_carries_the_speaker(self, fact_store):
-        stored = await store_extracted_facts(
-            fact_store,
-            facts=[
-                {"subject": "alice", "predicate": "prefers", "object": "tea"},
-                {"subject": "alice", "predicate": "works_at", "object": "acme"},
-            ],
-            source_interaction_id="i-1", asserted_at=1_100.0,
-            session_id="legacy", speaker_id="amber-lynx",
-        )
-
-        assert stored == 2
-        assert await _fact_speakers(fact_store) == {"amber-lynx"}
-
-    async def test_a_speakerless_source_is_null(self, fact_store):
-        await store_extracted_facts(
-            fact_store,
-            facts=[{"subject": "alice", "predicate": "prefers", "object": "tea"}],
-            source_interaction_id="i-1", asserted_at=1_100.0,
-            session_id="legacy", speaker_id=None,
-        )
-
-        assert await _fact_speakers(fact_store) == {None}
-
-    async def test_speaker_is_not_the_subject(self, fact_store):
-        """The two columns answer different questions — who SAID it
-        versus who it is ABOUT — and a counterparty fact differs in
-        both."""
-        await store_extracted_facts(
-            fact_store,
-            facts=[{"subject": "alice", "predicate": "prefers", "object": "tea"}],
-            source_interaction_id="i-1", asserted_at=1_100.0,
-            session_id="legacy", speaker_id="amber-lynx",
-        )
-
-        db = fact_store._ensure_db()
-        async with db.execute("SELECT subject, speaker_id FROM facts") as cur:
-            rows = await cur.fetchall()
-        assert [tuple(r) for r in rows] == [("alice", "amber-lynx")]
-
-
 async def _noop() -> None:
     return None
 
@@ -308,20 +263,6 @@ async def _episode_speaker(memory, interaction_id: str) -> str | None:
         row = await cursor.fetchone()
     assert row is not None, "the close path wrote no episode"
     return row[0]
-
-
-async def _fact_speakers(fact_store: FactStore) -> set[str | None]:
-    db = fact_store._ensure_db()
-    async with db.execute("SELECT speaker_id FROM facts") as cursor:
-        return {row[0] for row in await cursor.fetchall()}
-
-
-@pytest.fixture
-async def fact_store():
-    store = FactStore(agent_id="test-agent", db_path=":memory:")
-    await store.initialize()
-    yield store
-    await store.close()
 
 
 class TestAllExcludedDegradesCheaply:
@@ -397,50 +338,6 @@ class TestThePromptCountsWhatItShows:
         prompt = llm.create_message.call_args.kwargs["messages"][0]["content"]
         assert "Turns: 1\n" in prompt
         assert "iron-fox" not in prompt
-
-
-class TestTheFactDispatchCarriesItThrough:
-    """The Phase-2 seam, not just the leaf.
-
-    ``TestFactProjection`` above calls ``store_extracted_facts`` with an
-    explicit speaker, so it pins the WRITE but not the WIRING: delete
-    ``dispatch_facts_from_response``'s ``speaker_id=`` argument and every
-    close-extracted fact silently reverts to NULL with that class still
-    green.  This is the same standard
-    ``test_interaction_classification_capture`` sets for the RFC 0037 §C
-    capture, whose docstring calls the dispatch pass-through "the only
-    coverage" of that seam — the speaker projection rides the same
-    function and needs the same pin.
-    """
-
-    async def test_the_records_speaker_reaches_the_stored_tuples(
-        self, fact_store,
-    ):
-        await dispatch_facts_from_response(
-            fact_store=fact_store,
-            facts_raw=json.dumps([
-                {"subject": "alice", "predicate": "prefers", "object": "tea"},
-                {"subject": "alice", "predicate": "works_at", "object": "acme"},
-            ]),
-            interaction=_record("amber-lynx"),
-            agent_id="test-agent",
-            session_id="legacy",
-        )
-
-        assert await _fact_speakers(fact_store) == {"amber-lynx"}
-
-    async def test_a_speakerless_record_dispatches_null(self, fact_store):
-        await dispatch_facts_from_response(
-            fact_store=fact_store,
-            facts_raw=json.dumps([
-                {"subject": "alice", "predicate": "prefers", "object": "tea"},
-            ]),
-            interaction=_record(""),
-            agent_id="test-agent",
-            session_id="legacy",
-        )
-
-        assert await _fact_speakers(fact_store) == {None}
 
 
 class _SingleTurnHarness(_EpisodeRoutingMixin):
